% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlenewschaRt.R
\name{googlenewschaRt}
\alias{googlenewschaRt}
\title{Puerto Rico Google News Chart Making}
\usage{
googlenewschaRt(filename = "R_News_1NOV17.csv")
}
\arguments{
\item{filename}{"Puerto Rico 23 Sept - 26 OCT best.csv"}
}
\description{
googlenewschaRt: This function takes the news data frame and makes the sentiment chart
}
\examples{
googlenewschaRt
}
\keyword{Puerto}
\keyword{Rico}
