% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puertoRicoR.R
\name{helpme}
\alias{helpme}
\title{helper function}
\usage{
helpme(filename = "Puerto Rico 23 Sept - 26 OCT best.csv")
}
\arguments{
\item{filename}{"Puerto Rico 23 Sept - 26 OCT best.csv"}
}
\description{
helper:
}
\examples{
helper()
}
\keyword{Puerto}
\keyword{Rico}
