% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataanalysispR.R
\name{helpmes}
\alias{helpmes}
\title{this helper function}
\usage{
helpmes(filename = "best.csv")
}
\arguments{
\item{filename}{"Puerto Rico 23 Sept - 26 OCT best.csv"}
}
\description{
helpers:
}
\examples{
helpers()
}
\keyword{Puerto}
\keyword{Ricos}
