% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newsscRape.R
\name{newsscRape}
\alias{newsscRape}
\title{update news}
\usage{
newsscRape(oldfile = "R_News_25OCT17a.csv", newfile = "R_News_27OCT17a.csv")
}
\description{
updatenews: This function updates your news search for "Puerto Rico" and saves it to the working directory
}
\examples{
newspull
}
\keyword{Google}
\keyword{News}
