% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataanalysispR.R, R/puertoRicoR.R
\name{pRanalysis}
\alias{pRanalysis}
\title{Puerto Rico Function}
\usage{
pRanalysis(filename = "Puerto Rico 23 Sept - 26 OCT best.csv")

pRanalysis(filename = "Puerto Rico 23 Sept - 26 OCT best.csv")
}
\arguments{
\item{filename}{"Name of file to lookup -- csv"}

\item{filename}{"Puerto Rico 23 Sept - 26 OCT best.csv"}
}
\description{
This function takes a twitter dataframe and produces the twitter sentiment graph over time

pRanalysis: This function takes a twitter dataframe and produces the twitter sentiment graph over time
}
\examples{
pRanalysis
pRtwitter()
}
\keyword{Puerto}
\keyword{Rico}
