% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puertoRicoR.R
\name{twbarplot}
\alias{twbarplot}
\title{bar plot common words}
\usage{
twbarplot(filename = "best.csv")
}
\arguments{
\item{filename}{"best.csv"}
}
\description{
common words barplot:
}
\examples{
twbarplot()
}
\keyword{Puerto}
\keyword{Rico}
