% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRanalysis.R
\name{pRanalysis}
\alias{pRanalysis}
\title{Puerto Rico Sentiment Tracking Function}
\usage{
pRanalysis(filename = "Puerto Rico 23 Sept - 26 OCT best.csv")
}
\arguments{
\item{filename}{"Puerto Rico 23 Sept - 26 OCT best.csv"}
}
\description{
pRanalysis: This function takes a twitter dataframe and produces the twitter sentiment graph over time
}
\examples{
pRtwitter()
}
\keyword{Puerto}
\keyword{Rico}
